split Product in subcomponents
